refactor(cli): split command groups into modules
