+++ src/gispulse/cli.py
@@ -19,12 +19,8 @@ Usage::
 from __future__ import annotations
 
 from pathlib import Path
-from typing import TYPE_CHECKING, Optional
 
 import typer
-
-if TYPE_CHECKING:
-    from packaging.version import Version
 
 app = typer.Typer(
     name="gispulse",
@@ -675,466 +671,14 @@ def engine(
 
 
 # ---------------------------------------------------------------------------
-# jobs sub-group
+# jobs and update command groups
 # ---------------------------------------------------------------------------
 
-jobs_app = typer.Typer(
-    name="jobs",
-    help="Manage GISPulse jobs (list, status, cancel).",
-    add_completion=False,
-)
+from gispulse.cli_jobs import jobs_app  # noqa: E402
+from gispulse.cli_update import cmd_update, startup_update_check  # noqa: E402
+
 app.add_typer(jobs_app)
-
-
-def _jobs_http(host: str, api_key: str | None):
-    """Return an httpx.Client configured for the given host."""
-    import httpx
-
-    headers = {}
-    if api_key:
-        headers["X-API-Key"] = api_key
-    return httpx.Client(base_url=host.rstrip("/"), headers=headers, timeout=10.0)
-
-
-@jobs_app.command("list")
-def jobs_list(
-    host: str = typer.Option(
-        "http://localhost:8001",
-        "--host",
-        "-H",
-        help="GISPulse API base URL.",
-        envvar="GISPULSE_HOST",
-    ),
-    api_key: str | None = typer.Option(
-        None,
-        "--api-key",
-        help="API key for authentication.",
-        envvar="GISPULSE_API_KEY",
-    ),
-    limit: int = typer.Option(50, "--limit", "-n", help="Maximum number of jobs to show."),
-) -> None:
-    """List recent jobs."""
-    import httpx
-
-    with _jobs_http(host, api_key) as http:
-        try:
-            resp = http.get("/jobs", params={"limit": limit})
-            resp.raise_for_status()
-        except httpx.HTTPStatusError as exc:
-            typer.echo(f"Error {exc.response.status_code}: {exc.response.text}", err=True)
-            raise typer.Exit(1)
-        except httpx.RequestError as exc:
-            typer.echo(f"Connection error: {exc}", err=True)
-            raise typer.Exit(1)
-
-    data = resp.json()
-    items = data if isinstance(data, list) else data.get("items", [])
-
-    if not items:
-        typer.echo("No jobs found.")
-        return
-
-    header = f"{'ID':<38}  {'STATUS':<10}  {'NAME':<24}  {'ATTEMPTS':>8}  {'DURATION':>10}"
-    typer.echo(header)
-    typer.echo("-" * len(header))
-    for job in items:
-        duration = job.get("duration_seconds")
-        dur_str = f"{duration:.1f}s" if duration is not None else "—"
-        typer.echo(
-            f"{job['id']:<38}  {job['status']:<10}  {job.get('name', ''):<24}"
-            f"  {job.get('attempts', 0):>8}  {dur_str:>10}"
-        )
-
-
-@jobs_app.command("status")
-def jobs_status(
-    job_id: str = typer.Argument(..., help="Job UUID."),
-    host: str = typer.Option(
-        "http://localhost:8001",
-        "--host",
-        "-H",
-        envvar="GISPULSE_HOST",
-    ),
-    api_key: str | None = typer.Option(
-        None,
-        "--api-key",
-        envvar="GISPULSE_API_KEY",
-    ),
-) -> None:
-    """Show detailed status of a job."""
-    import httpx
-
-    with _jobs_http(host, api_key) as http:
-        try:
-            resp = http.get(f"/jobs/{job_id}")
-            if resp.status_code == 404:
-                typer.echo(f"Job '{job_id}' not found.", err=True)
-                raise typer.Exit(1)
-            resp.raise_for_status()
-        except httpx.HTTPStatusError as exc:
-            typer.echo(f"Error {exc.response.status_code}: {exc.response.text}", err=True)
-            raise typer.Exit(1)
-        except httpx.RequestError as exc:
-            typer.echo(f"Connection error: {exc}", err=True)
-            raise typer.Exit(1)
-
-    job = resp.json()
-    typer.echo(f"ID:        {job['id']}")
-    typer.echo(f"Name:      {job.get('name', '')}")
-    typer.echo(f"Status:    {job['status']}")
-    typer.echo(f"Attempts:  {job.get('attempts', 0)}")
-    if job.get("dataset_id"):
-        typer.echo(f"Dataset:   {job['dataset_id']}")
-    if job.get("started_at"):
-        typer.echo(f"Started:   {job['started_at']}")
-    if job.get("completed_at"):
-        typer.echo(f"Completed: {job['completed_at']}")
-    if job.get("duration_seconds") is not None:
-        typer.echo(f"Duration:  {job['duration_seconds']:.1f}s")
-    if job.get("error_message"):
-        typer.echo(f"Error:     {job['error_message']}")
-    if job.get("result_path"):
-        typer.echo(f"Result:    {job['result_path']}")
-
-
-@jobs_app.command("cancel")
-def jobs_cancel(
-    job_id: str = typer.Argument(..., help="Job UUID."),
-    host: str = typer.Option(
-        "http://localhost:8001",
-        "--host",
-        "-H",
-        envvar="GISPULSE_HOST",
-    ),
-    api_key: str | None = typer.Option(
-        None,
-        "--api-key",
-        envvar="GISPULSE_API_KEY",
-    ),
-) -> None:
-    """Cancel a pending or running job."""
-    import httpx
-
-    with _jobs_http(host, api_key) as http:
-        try:
-            resp = http.post(f"/jobs/{job_id}/cancel")
-            if resp.status_code == 404:
-                typer.echo(f"Job '{job_id}' not found.", err=True)
-                raise typer.Exit(1)
-            if resp.status_code == 409:
-                detail = resp.json().get("detail", "")
-                typer.echo(f"Cannot cancel: {detail}", err=True)
-                raise typer.Exit(1)
-            resp.raise_for_status()
-        except httpx.HTTPStatusError as exc:
-            typer.echo(f"Error {exc.response.status_code}: {exc.response.text}", err=True)
-            raise typer.Exit(1)
-        except httpx.RequestError as exc:
-            typer.echo(f"Connection error: {exc}", err=True)
-            raise typer.Exit(1)
-
-    job = resp.json()
-    typer.echo(f"Job '{job_id}' cancelled (status: {job['status']}).")
-
-
-# ---------------------------------------------------------------------------
-# update command + startup check
-# ---------------------------------------------------------------------------
-
-_GITHUB_RELEASES_URL = "https://api.github.com/repos/imagodata/gispulse/releases/latest"
-_UPDATE_CHECK_CACHE = Path("~/.gispulse/update-check.json").expanduser()
-_UPDATE_CHECK_INTERVAL_SECONDS = 86400  # 24 h
-
-
-def _get_installed_version() -> str:
-    """Return the installed version of gispulse."""
-    from importlib.metadata import version as pkg_version
-
-    try:
-        return pkg_version("gispulse")
-    except Exception:
-        return "0.1.0"
-
-
-def _fetch_latest_release() -> dict | None:
-    """Fetch latest release info from GitHub API. Returns None on failure."""
-    import json
-    import urllib.request
-
-    req = urllib.request.Request(
-        _GITHUB_RELEASES_URL,
-        headers={"Accept": "application/vnd.github.v3+json", "User-Agent": "gispulse-cli"},
-    )
-    try:
-        with urllib.request.urlopen(req, timeout=10) as resp:
-            return json.loads(resp.read().decode())
-    except Exception:
-        return None
-
-
-def _parse_version(v: str) -> Version:
-    """Parse a version string, stripping leading 'v' if present."""
-    from packaging.version import Version
-
-    return Version(v.lstrip("v"))
-
-
-def _detect_install_mode() -> str:
-    """Detect how gispulse was installed: 'pip', 'homebrew', or 'binary'."""
-    import shutil
-    import subprocess
-    import sys
-
-    # pip: running from a site-packages environment
-    if "site-packages" in (sys.executable or ""):
-        return "pip"
-
-    # homebrew: brew is available and gispulse is in its list
-    brew = shutil.which("brew")
-    if brew:
-        try:
-            result = subprocess.run(
-                [brew, "list", "--formula"],
-                capture_output=True,
-                text=True,
-                timeout=10,
-            )
-            if "gispulse" in result.stdout.split():
-                return "homebrew"
-        except Exception:
-            pass
-
-    return "binary"
-
-
-def _truncate_changelog(body: str | None, max_lines: int = 12) -> str:
-    """Truncate release body for display."""
-    if not body:
-        return "(no changelog)"
-    lines = body.strip().splitlines()
-    if len(lines) <= max_lines:
-        return body.strip()
-    return "\n".join(lines[:max_lines]) + f"\n  ... ({len(lines) - max_lines} more lines)"
-
-
-@app.command()
-def update(
-    check: bool = typer.Option(False, "--check", help="Check only, do not install."),
-    force: bool = typer.Option(False, "--force", help="Update even if already at latest version."),
-) -> None:
-    """Check for updates and self-update GISPulse."""
-    import subprocess
-    import shutil
-    import sys
-    import tempfile
-
-    current = _get_installed_version()
-    typer.echo(f"Current version: v{current}")
-
-    release = _fetch_latest_release()
-    if release is None:
-        typer.echo("Error: could not reach GitHub API (no network or rate-limited).", err=True)
-        raise typer.Exit(1)
-
-    tag = release.get("tag_name", "")
-    if not tag:
-        typer.echo("Error: no tag found in latest release.", err=True)
-        raise typer.Exit(1)
-
-    try:
-        current_v = _parse_version(current)
-        latest_v = _parse_version(tag)
-    except Exception as e:
-        typer.echo(f"Error parsing versions: {e}", err=True)
-        raise typer.Exit(1)
-
-    typer.echo(f"Latest version:  v{latest_v}")
-
-    is_outdated = latest_v > current_v
-
-    if check:
-        if is_outdated:
-            typer.echo(f"\nUpdate available: v{current} -> v{latest_v}")
-            typer.echo("Run `gispulse update` to upgrade.")
-            raise typer.Exit(1)
-        else:
-            typer.echo(f"\nGISPulse v{current} is up to date.")
-            raise typer.Exit(0)
-
-    if not is_outdated and not force:
-        typer.echo(f"\nGISPulse v{current} is up to date.")
-        return
-
-    # Show changelog
-    changelog = _truncate_changelog(release.get("body"))
-    typer.echo(f"\nChangelog:\n  {changelog.replace(chr(10), chr(10) + '  ')}")
-
-    # Confirm unless --force
-    if not force:
-        confirm = typer.confirm(f"\nUpgrade to v{latest_v}?")
-        if not confirm:
-            typer.echo("Cancelled.")
-            raise typer.Exit(0)
-
-    mode = _detect_install_mode()
-    typer.echo(f"\nInstall mode: {mode}")
-
-    if mode == "pip":
-        typer.echo(f"Running: pip install --upgrade gispulse=={latest_v}")
-        result = subprocess.run(
-            [sys.executable, "-m", "pip", "install", "--upgrade", f"gispulse=={latest_v}"],
-            capture_output=True,
-            text=True,
-        )
-        if result.returncode != 0:
-            typer.echo(f"pip upgrade failed:\n{result.stderr}", err=True)
-            raise typer.Exit(1)
-        typer.echo(f"Successfully updated to v{latest_v}.")
-
-    elif mode == "homebrew":
-        brew = shutil.which("brew")
-        typer.echo("Running: brew upgrade gispulse")
-        result = subprocess.run(
-            [brew, "upgrade", "gispulse"],  # type: ignore[arg-type]
-            capture_output=True,
-            text=True,
-        )
-        if result.returncode != 0:
-            typer.echo(f"brew upgrade failed:\n{result.stderr}", err=True)
-            raise typer.Exit(1)
-        typer.echo(f"Successfully updated to v{latest_v}.")
-
-    elif mode == "binary":
-        # Find a matching asset for the current platform
-        import platform
-        import urllib.request
-
-        system = platform.system().lower()  # linux, darwin, windows
-        machine = platform.machine().lower()  # x86_64, arm64, aarch64
-
-        assets = release.get("assets", [])
-        matching = [
-            a for a in assets
-            if system in a["name"].lower() and (machine in a["name"].lower()
-                                                  or ("amd64" in a["name"].lower() and machine == "x86_64")
-                                                  or ("arm64" in a["name"].lower() and machine == "aarch64"))
-        ]
-
-        if not matching:
-            typer.echo(
-                f"Error: no binary asset found for {system}/{machine} in release v{latest_v}.\n"
-                f"Available assets: {[a['name'] for a in assets]}",
-                err=True,
-            )
-            typer.echo("Try: pip install --upgrade gispulse", err=True)
-            raise typer.Exit(1)
-
-        asset = matching[0]
-        download_url = asset["browser_download_url"]
-        typer.echo(f"Downloading {asset['name']} ...")
-
-        try:
-            with tempfile.NamedTemporaryFile(delete=False, suffix=Path(asset["name"]).suffix) as tmp:
-                req = urllib.request.Request(download_url, headers={"User-Agent": "gispulse-cli"})
-                with urllib.request.urlopen(req, timeout=120) as resp:
-                    tmp.write(resp.read())
-                tmp_path = Path(tmp.name)
-        except Exception as e:
-            typer.echo(f"Download failed: {e}", err=True)
-            raise typer.Exit(1)
-
-        # Replace current binary
-        current_bin = Path(shutil.which("gispulse") or sys.executable)
-        try:
-            import stat
-
-            tmp_path.chmod(current_bin.stat().st_mode | stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH)
-            backup = current_bin.with_suffix(".bak")
-            shutil.copy2(current_bin, backup)
-            shutil.move(str(tmp_path), str(current_bin))
-            typer.echo(f"Successfully updated to v{latest_v}.")
-            typer.echo(f"Backup saved as {backup}")
-        except PermissionError:
-            typer.echo(
-                f"Permission denied. Try:\n  sudo cp {tmp_path} {current_bin}",
-                err=True,
-            )
-            raise typer.Exit(1)
-
-    # Update the cache so startup check doesn't nag
-    _write_update_cache(str(latest_v))
-
-
-def _write_update_cache(latest_version: str) -> None:
-    """Write update check result to cache file."""
-    import json
-    import time
-
-    _UPDATE_CHECK_CACHE.parent.mkdir(parents=True, exist_ok=True)
-    cache = {
-        "checked_at": time.time(),
-        "latest_version": latest_version,
-    }
-    try:
-        _UPDATE_CHECK_CACHE.write_text(json.dumps(cache))
-    except OSError:
-        pass  # non-critical
-
-
-def _read_update_cache() -> dict | None:
-    """Read cached update check result. Returns None if stale or missing."""
-    import json
-    import time
-
-    if not _UPDATE_CHECK_CACHE.exists():
-        return None
-    try:
-        cache = json.loads(_UPDATE_CHECK_CACHE.read_text())
-        if time.time() - cache.get("checked_at", 0) > _UPDATE_CHECK_INTERVAL_SECONDS:
-            return None
-        return cache
-    except (json.JSONDecodeError, OSError, KeyError):
-        return None
-
-
-def _startup_update_check() -> None:
-    """Non-blocking update check at CLI startup. Prints a notice if outdated.
-
-    Respects GISPULSE_NO_UPDATE_CHECK=1 env var.
-    Caches results for 24h in ~/.gispulse/update-check.json.
-    """
-
-    from gispulse.core.config import settings as _cfg3
-    if _cfg3.telemetry.no_update_check:
-        return
-
-    # Check cache first
-    cache = _read_update_cache()
-    if cache is None:
-        # Perform the check (synchronous but fast — single HTTP call)
-        release = _fetch_latest_release()
-        if release is None:
-            return  # no network, silently skip
-        tag = release.get("tag_name", "")
-        if not tag:
-            return
-        latest_str = tag.lstrip("v")
-        _write_update_cache(latest_str)
-    else:
-        latest_str = cache["latest_version"]
-
-    try:
-        current_v = _parse_version(_get_installed_version())
-        latest_v = _parse_version(latest_str)
-    except Exception:
-        return
-
-    if latest_v > current_v:
-        typer.echo(
-            f"\nA new version of GISPulse is available: v{latest_v} (current: v{current_v}).\n"
-            f"Run `gispulse update` to upgrade.\n",
-            err=True,
-        )
+app.command(name="update")(cmd_update)
 
 
 # Register the startup check as a Typer callback
@@ -1157,7 +701,7 @@ def _cli_callback(
         raise typer.Exit(0)
     # Don't run update check for the 'update' command itself
     if ctx.invoked_subcommand != "update":
-        _startup_update_check()
+        startup_update_check()
 
 
 # ---------------------------------------------------------------------------
@@ -1202,305 +746,14 @@ def telemetry_callback(
 
 
 # ---------------------------------------------------------------------------
-# Marketplace sub-commands
+# marketplace and template sub-groups
 # ---------------------------------------------------------------------------
-marketplace_app = typer.Typer(
-    name="marketplace",
-    help="Manage GISPulse capability plugins.",
-    add_completion=False,
-)
+
+from gispulse.cli_marketplace import marketplace_app  # noqa: E402
+from gispulse.cli_templates import template_app  # noqa: E402
+
 app.add_typer(marketplace_app, name="marketplace")
-
-_PLUGIN_PREFIX = "gispulse-cap-"
-_REGISTRY_URL = (
-    "https://raw.githubusercontent.com/gispulse/marketplace/main/registry.json"
-)
-
-
-@marketplace_app.command("list")
-def marketplace_list() -> None:
-    """List installed GISPulse plugins (entry-point based)."""
-    from gispulse.capabilities.registry import list_plugins
-
-    plugins = list_plugins()
-    if not plugins:
-        typer.echo("No plugins installed.")
-        typer.echo(
-            "\nInstall one with:  gispulse marketplace install <name>"
-        )
-        return
-
-    typer.echo(f"{len(plugins)} plugin(s) installed:\n")
-    for p in plugins:
-        typer.echo(f"  - {p['name']}  ({p['module']})")
-
-
-@marketplace_app.command("search")
-def marketplace_search(
-    query: str = typer.Argument(..., help="Search term (e.g. 'ftth', 'raster')."),
-) -> None:
-    """Search PyPI for GISPulse capability packages."""
-    import json
-    import urllib.request
-    import urllib.error
-
-    url = "https://pypi.org/simple/"
-    typer.echo(f"Searching PyPI for '{_PLUGIN_PREFIX}*' matching '{query}'...")
-
-    # Use PyPI JSON API to search by project name pattern
-    search_url = "https://pypi.org/simple/"
-    try:
-        req = urllib.request.Request(search_url, headers={"Accept": "application/vnd.pypi.simple.v1+json"})
-        with urllib.request.urlopen(req, timeout=10) as resp:
-            data = json.loads(resp.read())
-            projects = data.get("projects", [])
-            matches = [
-                p["name"]
-                for p in projects
-                if p["name"].startswith(_PLUGIN_PREFIX) and query.lower() in p["name"].lower()
-            ]
-    except Exception:
-        # Fallback: try the curated registry
-        matches = []
-        try:
-            req = urllib.request.Request(_REGISTRY_URL)
-            with urllib.request.urlopen(req, timeout=10) as resp:
-                registry = json.loads(resp.read())
-                matches = [
-                    p["package"]
-                    for p in registry.get("plugins", [])
-                    if query.lower() in p.get("name", "").lower()
-                    or query.lower() in p.get("description", "").lower()
-                ]
-        except Exception:
-            typer.echo("Error: could not reach PyPI or plugin registry.", err=True)
-            raise typer.Exit(1)
-
-    if not matches:
-        typer.echo("No matching plugins found.")
-        return
-
-    typer.echo(f"\n{len(matches)} result(s):\n")
-    for name in matches:
-        typer.echo(f"  - {name}")
-    typer.echo("\nInstall with:  gispulse marketplace install <name>")
-
-
-@marketplace_app.command("install")
-def marketplace_install(
-    name: str = typer.Argument(
-        ..., help="Plugin name (e.g. 'ftth'). Will install gispulse-cap-<name>."
-    ),
-    upgrade: bool = typer.Option(False, "--upgrade", "-U", help="Upgrade if already installed."),
-) -> None:
-    """Install a GISPulse capability plugin from PyPI."""
-    import subprocess
-    import sys
-
-    package = f"{_PLUGIN_PREFIX}{name}" if not name.startswith(_PLUGIN_PREFIX) else name
-    cmd = [sys.executable, "-m", "pip", "install", package]
-    if upgrade:
-        cmd.append("--upgrade")
-
-    typer.echo(f"Installing {package}...")
-    result = subprocess.run(cmd, capture_output=True, text=True)
-
-    if result.returncode != 0:
-        typer.echo(f"Error installing {package}:\n{result.stderr}", err=True)
-        raise typer.Exit(1)
-
-    typer.echo(f"Installed {package} successfully.")
-    typer.echo("Restart GISPulse to use the new capability.")
-
-
-@marketplace_app.command("uninstall")
-def marketplace_uninstall(
-    name: str = typer.Argument(
-        ..., help="Plugin name (e.g. 'ftth'). Will uninstall gispulse-cap-<name>."
-    ),
-) -> None:
-    """Uninstall a GISPulse capability plugin."""
-    import subprocess
-    import sys
-
-    package = f"{_PLUGIN_PREFIX}{name}" if not name.startswith(_PLUGIN_PREFIX) else name
-
-    typer.echo(f"Uninstalling {package}...")
-    result = subprocess.run(
-        [sys.executable, "-m", "pip", "uninstall", "-y", package],
-        capture_output=True,
-        text=True,
-    )
-
-    if result.returncode != 0:
-        typer.echo(f"Error uninstalling {package}:\n{result.stderr}", err=True)
-        raise typer.Exit(1)
-
-    typer.echo(f"Uninstalled {package}.")
-
-
-@marketplace_app.command("info")
-def marketplace_info(
-    name: str = typer.Argument(
-        ..., help="Plugin name (e.g. 'ftth')."
-    ),
-) -> None:
-    """Show details about an installed plugin."""
-    from importlib.metadata import PackageNotFoundError
-
-    package = f"{_PLUGIN_PREFIX}{name}" if not name.startswith(_PLUGIN_PREFIX) else name
-
-    try:
-        from importlib.metadata import metadata as pkg_metadata
-
-        meta = pkg_metadata(package)
-        typer.echo(f"Package:     {meta['Name']}")
-        typer.echo(f"Version:     {meta['Version']}")
-        typer.echo(f"Summary:     {meta.get('Summary', 'N/A')}")
-        typer.echo(f"Author:      {meta.get('Author', meta.get('Author-email', 'N/A'))}")
-        typer.echo(f"License:     {meta.get('License', 'N/A')}")
-        typer.echo(f"Home-page:   {meta.get('Home-page', 'N/A')}")
-    except PackageNotFoundError:
-        typer.echo(f"Package '{package}' is not installed.", err=True)
-        raise typer.Exit(1)
-
-
-# ---------------------------------------------------------------------------
-# template sub-group
-# ---------------------------------------------------------------------------
-
-template_app = typer.Typer(
-    name="template",
-    help="Manage and use GISPulse pipeline templates.",
-    add_completion=False,
-)
 app.add_typer(template_app)
-
-#: Directory where built-in templates are stored (relative to this file's package root).
-_TEMPLATES_DIR = Path(__file__).resolve().parents[2] / "templates"
-
-
-def _load_template_index() -> list[dict]:
-    """Return metadata for every .json template in the templates directory."""
-    import json
-
-    if not _TEMPLATES_DIR.exists():
-        return []
-
-    entries: list[dict] = []
-    for tpl_path in sorted(_TEMPLATES_DIR.glob("*.json")):
-        try:
-            data = json.loads(tpl_path.read_text(encoding="utf-8"))
-            steps = len(data) if isinstance(data, list) else 1
-            entries.append(
-                {
-                    "name": tpl_path.stem,
-                    "path": tpl_path,
-                    "steps": steps,
-                    "capabilities": (
-                        sorted({r.get("capability", "?") for r in data})
-                        if isinstance(data, list)
-                        else []
-                    ),
-                }
-            )
-        except Exception:
-            entries.append({"name": tpl_path.stem, "path": tpl_path, "steps": 0, "capabilities": []})
-
-    return entries
-
-
-@template_app.command("list")
-def template_list() -> None:
-    """List available built-in pipeline templates."""
-    entries = _load_template_index()
-
-    if not entries:
-        typer.echo(
-            f"No templates found in {_TEMPLATES_DIR}. "
-            "Re-install GISPulse or run from the project root."
-        )
-        return
-
-    typer.echo(f"Available templates ({len(entries)}):\n")
-    typer.echo(f"  {'Name':<35} {'Steps':>5}  Capabilities")
-    typer.echo(f"  {'─' * 35} {'─' * 5}  {'─' * 40}")
-    for entry in entries:
-        caps = ", ".join(entry["capabilities"]) or "—"
-        typer.echo(f"  {entry['name']:<35} {entry['steps']:>5}  {caps}")
-
-    typer.echo("\nUse: gispulse template use <name> [-o <dest>]")
-
-
-@template_app.command("use")
-def template_use(
-    name: str = typer.Argument(..., help="Template name (without .json extension)."),
-    output: Path = typer.Option(
-        None,
-        "--output",
-        "-o",
-        help="Destination path (default: <name>.json in the current directory).",
-    ),
-) -> None:
-    """Copy a built-in pipeline template to the current directory (or a given path).
-
-    Example::
-
-        gispulse template use validation_plu_cnig
-        gispulse template use ftth_network_analysis -o rules/ftth.json
-    """
-    import shutil
-
-    # Resolve template source
-    tpl_path = _TEMPLATES_DIR / f"{name}.json"
-    if not tpl_path.exists():
-        # Try exact match (user passed name.json)
-        tpl_path = _TEMPLATES_DIR / name
-    if not tpl_path.exists():
-        entries = _load_template_index()
-        available = [e["name"] for e in entries]
-        typer.echo(
-            f"Error: template '{name}' not found.\n"
-            f"Available: {', '.join(available) or 'none'}",
-            err=True,
-        )
-        raise typer.Exit(1)
-
-    dest = output or Path.cwd() / tpl_path.name
-    if dest.exists():
-        typer.echo(f"Warning: {dest} already exists — overwriting.", err=True)
-
-    dest.parent.mkdir(parents=True, exist_ok=True)
-    shutil.copy2(tpl_path, dest)
-    typer.echo(f"Template '{name}' copied to {dest}")
-    typer.echo(f"Edit the file then run: gispulse run <input> --rules {dest} -o output.gpkg")
-
-
-@template_app.command("workflow")
-def workflow_execute(
-    name: str = typer.Argument(..., help="Workflow name (ex: ftth_network_analysis)"),
-    input_path: Path = typer.Argument(..., help="Chemin vers le GPKG d'entrée"),
-    output_path: Optional[Path] = typer.Option(None, "--output", "-o", help="Chemin de sortie (GPKG)"),
-) -> None:
-    """Exécute un workflow intégré (ex: ftth_network_analysis).
-
-    Example::
-
-        gispulse template workflow ftth_network_analysis input.gpkg -o output.gpkg
-    """
-    from gispulse.workflows.ftth_network_analysis import FTTHNetworkAnalysisWorkflow
-    from gispulse.persistence.gpkg import read_gpkg
-
-    dataset = read_gpkg(input_path)
-    template_path = _TEMPLATES_DIR / f"{name}.json"
-    
-    if not template_path.exists():
-        typer.echo(f"Error: workflow template '{name}' not found.", err=True)
-        raise typer.Exit(1)
-
-    workflow = FTTHNetworkAnalysisWorkflow(template_path, dataset)
-    result = workflow.run(output_path)
-    typer.echo(f"Workflow '{name}' terminé. Résultat: {len(result.layers)} couches.")
 
 
 # ---------------------------------------------------------------------------
